+++ CHANGELOG.md
@@ -31,11 +31,15 @@ Changelog
 
 ### 1.12.0
 
-Fixes https://github.com/lasso-js/lasso-minify-js/issues/1 - Upgrade to the latest version of UglifyJS
+- Fixes https://github.com/lasso-js/lasso-minify-js/issues/1 - Upgrade to the latest version of UglifyJS
+
+### 1.12.3
+
+- Adds support for `noConflict` lasso configuration option
 
 ### 1.11.12
 
-Fixes [#82](https://github.com/lasso-js/lasso/issues/82) - Make lasso a true singleton
+- Fixes [#82](https://github.com/lasso-js/lasso/issues/82) - Make lasso a true singleton
 
 ### 1.11.11